Upgrade Dockerfiles to go1.16 and alpine3.13
As part of upgrading codebase to golang v1.16, upgrade the
Dockerfile to use the golang oficial 1.16 alpine image, which
using alpine3.13 version.

After the previous commit, where the go.mod file was introduce,
there is not need to rely on the go `dep` tool for building the
dependencies. Thus, removing this from the docker image step.