only register containers that are in running and restarting state
During Sync() registrator is registering all docker containers including
containers that aren't running.
Afterwards it unregisters all containers again aren't in status created,
restarting, running or paused.

Change it to only register containers that are running or restarting.
It's not clear why containers are registered that are not running at
all.
This will allow remove the logic to mark containers as dying, the issue
that containers can be reregistered during shutdown can not happen
anymore.