remove tracking stopped containers as dying
Do not mark containers as dying after they have been stopped.
The mechanic causes that when a single container is restarted it is not
registered again. The containers is marked as dying until
ttl-dying-cleanup expired. If the container is restarted before the TTL
expires it ignored and not registered again.
Containers are also only unmarked as dying when bridge.add() is called,
if this did not happen despite the TTL expires and the container was
restarted it was also not registered again.
I guess the dying state was introduced to prevent that a container is
reregistered while it is being shutdown.
Because Sync() was changed to only consider containers in running state
it can not happen anymore.

The ttl-dying-cleanup config option is removed.