fix: docker events can get lost
A channel without buffer was used to receive dockerapi events.
Events are only sent to the channel if the operation does not block.
This means when registrator is still processing an event, events send to
the channel will get lost.

Make it unlikely to happen by using a channel with a buffer of 128 messages.